Drop `OpenRouterAPIError` in favor of httpx.HTTPStatusError.
Using processor-specific error binds the user specifically to OpenRouterModel. But the errors themselves are not backend-specific. The problem is - once we submit this pull request, people will start using it and chinging exception type would become a breaking change. At the same time we need to balance the time/effort one has to put to submit things. We don't want to block on devising "proper" exception hierarchy.

For this reason I think re-raising httpx.HTTPStatusError is a good middle ground: they provide a unified way to determine whether exception is retryable or not.